backend/requirements.txt
Django==2.2.3
django-templated-mail==1.1.1
djangorestframework==3.10.1
djangorestframework-jwt==1.8.0
djoser==1.7.0
mysqlclient==1.4.2.post1
npm==0.1.1
optional-django==0.1.0
pytz==2019.1
sqlparse==0.3.0
gunicorn==19.9.0
